+++ src/components/CodeEditor/CodeEditor.tsx
@@ -8,7 +8,7 @@ import { HTMLOptionsSetter, createRange, verify } from '../../utils';
 import { useAppDispatch, useAppSelector } from '../../store/store';
 import { htmlActions } from '../../store/sourceHtml/sourceHtml';
 import { EditorSelection } from '../../types/types';
-import {  themeSwitcher } from '../../utils/themeLoader';
+import {  themeSwitcher } from '../../utils';
 import { LS_MONACOTHEME } from '../../constants';
