+++ src/components/JSONEditor/JSONEditor.tsx
@@ -6,7 +6,7 @@ import { Editor, Monaco } from '@monaco-editor/react';
 import { useAppDispatch, useAppSelector } from '../../store/store';
 import { htmlActions } from '../../store/sourceHtml/sourceHtml';
 import { LS_MONACOTHEME } from '../../constants';
-import { themeSwitcher } from '../../utils/themeLoader';
+import { themeSwitcher } from '../../utils';
 
 export const JSONEditor: FC<{fontSize:number}> = ({ fontSize = 12 }) => {
   const json = useAppSelector((state) => state.htmlReducer.json);
