+++ src/components/index.ts
@@ -1,3 +1,7 @@
+export { ThemeSwitcher } from './themeSwitcher/ThemeSwitcher';
+export { Stats } from './Stats/Stats';
+export { JSONEditor } from './JSONEditor/JSONEditor';
+export { Images } from './Images/Images';
 export { Layout } from './Layout/Layout';
 export { Frame } from './Frame/Frame';
 export { CodeEditor } from './CodeEditor/CodeEditor';
