+++ src/main.tsx
@@ -13,7 +13,7 @@ import jsonWorker from 'monaco-editor/esm/vs/language/json/json.worker?worker';
 import htmlWorker from 'monaco-editor/esm/vs/language/html/html.worker?worker';
 import { Provider } from 'react-redux';
 import { store } from './store/store.ts';
-import { addRule } from './utils/linter.ts';
+import { addRule } from './utils';
 import { custom_snippets_emmet } from './constants.ts';
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
